# openapi.yaml
...
openapi: 3.1.0
info:
  title: aperIA - Python API
  description: '

    API do **aperIA** — ASPM (Application Security Posture Management) open-source.


    Um webhook de Pull Request do GitHub dispara um pipeline de scanners de

    segurança em 3 tiers (via Celery). O Claude raciocina como atacante

    (correlaciona findings, monta attack paths) e entrega patches como

    **GitHub code suggestions** para aprovação humana.


    > **Premissa inviolável:** o aperIA **nunca** aplica código sozinho.

    > Todo patch sai como code suggestion para aprovação humana.

    '
  contact:
    name: aperIA
  license:
    name: Open Source
  version: 1.0.0
paths:
...
  /github/connect:
    get:
      tags:
      - github
      summary: Iniciar conexão com o GitHub
      description: 'Gera a URL de instalação do App com um `state` assinado (10 min).


        O front-end deve redirecionar o usuário para `install_url`. Retorna 503

        enquanto o App não estiver configurado (`GITHUB_APP_SLUG` vazio).'
      operationId: connect_github_connect_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectResponse'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
      security:
      - HTTPBearer: []
...
  /github/repos:
    get:
      tags:
      - github
      summary: Listar repositórios disponíveis nas instalações do usuário
      description: 'Lista, ao vivo, os repositórios que as instalações do usuário enxergam.


        Para cada conta conectada, consulta o GitHub (installation token) e marca

        `active=True` nos repos que o usuário já ativou para análise. Falha ao

        consultar uma instalação é logada e ignorada (best-effort).'
      operationId: list_available_repos_github_repos_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AvailableRepo'
                type: array
                title: Response List Available Repos Github Repos Get
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
      security:
      - HTTPBearer: []
...
  /repositories:
    get:
      tags:
      - repositories
      summary: Listar repositórios ativados
      description: Lista os repositórios GitHub ativados para análise pelo usuário logado.
      operationId: list_repositories_repositories_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RepositoryResponse'
                type: array
                title: Response List Repositories Repositories Get
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
      security:
      - HTTPBearer: []
    post:
      tags:
      - repositories
      summary: Ativar repositório para análise
      description: 'Ativa um repositório GitHub para análise, vinculado a uma conta do usuário.


        A conta GitHub (`github_account_id`) precisa pertencer ao usuário logado;

        caso contrário retorna 404 (sem revelar se a conta existe para outro

        usuário).


        A operação é um **upsert** por `(user_id, github_repo_id)`: reativar um

        repositório desativado é um POST. A resposta traz o estado REAL persistido

        — em particular o `id` da linha existente, não o uuid gerado em memória

        (que não existiria no banco e faria `PATCH`/`DELETE` retornarem 404).


        `target_url` (opcional) é a URL onde a aplicação está publicada e vira o

        alvo do DAST no Tier 3. Ela passa pela validação anti-SSRF descrita em

        `PATCH /repositories/{id}` — URL malformada ou apontando para rede interna

        responde `422`. Por ser upsert, **omitir o campo preserva** a URL já

        gravada (senão um POST de reativação apagaria o alvo sem ninguém pedir);

        limpar é operação explícita do `PATCH` com `null`.'
      operationId: create_repository_repositories_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepositoryCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryResponse'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Conta GitHub nao encontrada.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /repositories/{repository_id}:
    get:
      tags:
      - repositories
      summary: Detalhe de um repositório
      description: Retorna o detalhe de um repositório. 404 se não pertencer ao usuário.
      operationId: get_repository_repositories__repository_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Repository Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryResponse'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Repositorio nao encontrado.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - repositories
      summary: Atualizar repositório (ativação e/ou alvo de DAST)
      description: "Atualização **parcial**: aplica só os campos presentes no corpo.\n\n`404` se o repositório\
        \ não existir ou não pertencer ao usuário (nunca 403).\n\nSemântica do corpo — `active` e `target_url`\
        \ são independentes:\n\n- campo ausente → não é tocado;\n- `{\"active\": false}` → desativa (contrato\
        \ inalterado, é o que o\n  front-end já envia);\n- `{\"target_url\": \"https://staging.acme.com\"\
        }` → define o alvo do DAST;\n- `{\"target_url\": null}` → **limpa** o alvo (o Tier 3 volta a pular\
        \ o ZAP).\n\n`null` em `target_url` é valor legítimo, não omissão — por isso a rota\ndecide por\
        \ presença da chave (`model_fields_set`), não por \"veio `None`\".\nCorpo vazio e `{\"active\"\
        : null}` são `422` explícitos, para que um erro de\ncliente não passe como no-op silencioso.\n\
        \n## Por que `target_url` é validada com uma blocklist de rede\n\nO `target_url` não é um link\
        \ exibido na UI: ele é entregue ao OWASP ZAP,\nque faz spider e **active scan** — dispara payloads\
        \ reais de SQLi, XSS e\npath traversal contra tudo que encontrar. Aceitar URL arbitrária tem duas\n\
        consequências diretas:\n\n1. **Ataque não autorizado a terceiros.** Cadastrar o site de outra\n\
        \   empresa transformaria o aperIA na origem de um ataque, com o IP da\n   nossa infraestrutura\
        \ no log da vítima.\n2. **SSRF privilegiado.** O worker de Tier 3 roda dentro da nossa rede e\n\
        \   alcança o que o usuário não alcança. `http://127.0.0.1:8090` é o\n   próprio ZAP; `http://10.0.0.5`\
        \ é banco/Redis; e\n   `http://169.254.169.254` é o endpoint de metadata de AWS/GCP/Azure, que\n\
        \   devolve **credenciais IAM temporárias da instância** — que voltariam ao\n   usuário dentro\
        \ dos findings do relatório.\n\nPor isso só passa URL http/https absoluta, sem credenciais embutidas,\
        \ com\nhost público: são recusados `localhost`, `127.0.0.0/8`, RFC1918\n(`10/8`, `172.16/12`,\
        \ `192.168/16`), link-local `169.254.0.0/16` (o\nmetadata incluso), CGNAT `100.64.0.0/10`, ULA/loopback\
        \ IPv6, IPv4 mapeado\nem IPv6, as formas decimal/hexadecimal de IP (`http://2130706433`),\nsufixos\
        \ internos (`.local`, `.internal`, `.localhost`) e hosts de rótulo\núnico (`http://zap`), que\
        \ só existem dentro da rede do worker.\n\nO que a validação **não** cobre, deliberadamente: um\
        \ domínio público cujo\nDNS aponta para dentro (rebinding). Resolver DNS no cadastro seria TOCTOU\n\
        — o registro muda entre validar e escanear — então a defesa desse vetor é\nde rede (egress policy\
        \ no container do ZAP), não de schema."
      operationId: update_repository_repositories__repository_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Repository Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepositoryUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryResponse'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Repositorio nao encontrado.
        '422':
          description: 'Payload invalido (nenhum campo, `active: null` ou `target_url` recusada).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
              examples:
                alvo_bloqueado:
                  summary: URL aponta para a rede interna
                  value:
                    detail:
                    - type: target_url_alvo_bloqueado
                      loc:
                      - body
                      - target_url
                      msg: 'Alvo bloqueado: 169.254.169.254 aponta para a rede interna. Informe a URL
                        publica do ambiente de staging/preview.'
                esquema_invalido:
                  summary: Esquema fora de http/https
                  value:
                    detail:
                    - type: target_url_esquema_invalido
                      loc:
                      - body
                      - target_url
                      msg: 'Esquema ''file'' nao e aceito: use http:// ou https://.'
    delete:
      tags:
      - repositories
      summary: Remover repositório
      description: Remove o vínculo do repositório. 404 se não pertencer ao usuário.
      operationId: delete_repository_repositories__repository_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Repository Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Repositorio nao encontrado.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /repositories/{repository_id}/scans:
    get:
      tags:
      - repositories
      summary: Listar scans do repositório
      description: Lista os scans (ScanJob) do repositório. 404 se não pertencer ao usuário.
      operationId: list_repository_scans_repositories__repository_id__scans_get
      security:
      - HTTPBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Repository Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanJobPage'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Repositorio nao encontrado.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AvailableRepo:
      properties:
        github_account_id:
          type: string
          format: uuid
          title: Github Account Id
        github_repo_id:
          type: integer
          title: Github Repo Id
        full_name:
          type: string
          title: Full Name
        url:
          type: string
          title: Url
        default_branch:
          type: string
          title: Default Branch
        active:
          type: boolean
          title: Active
        private:
          type: boolean
          title: Private
          default: false
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        pushed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Pushed At
      type: object
      required:
      - github_account_id
      - github_repo_id
      - full_name
      - url
      - default_branch
      - active
      title: AvailableRepo
      description: 'Repositório visível pela instalação, para a tela de ativação.


        ``active`` indica se já foi ativado para análise pelo usuário.

        ``github_account_id`` é o que o front envia depois em ``POST /repositories``.


        ``private``, ``language`` e ``pushed_at`` são metadados que já vêm no

        payload de ``GET /installation/repositories`` — expostos aqui para a tela

        de ativação não precisar de nenhuma chamada extra ao GitHub.'
    ConnectResponse:
      properties:
        install_url:
          type: string
          title: Install Url
      type: object
      required:
      - install_url
      title: ConnectResponse
      description: URL de instalação do GitHub App para onde o front deve redirecionar.
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...